fix(tests): breadcrumbs tests failing
Before this commit the tests were failling due to an operation inside
BreadCrumbs class. When it tries to call `self.addItem` the first
element of the list `path` was empty. This caused and error at least at
Linux machines because it's empty.

This commit fixes the issue by adding `os.path.sep` as a fallback.