feat(autosave): Add autosave functionality to texteditor
Before this commit the saving had to be done manually by the user.

This commit adds the autosaving behavior using tkinter method `after`
with a default timer of 10 seconds. The feature can be enable/disabled,
and the timer changed at config/__init__ file.

The saving its done for all the active files.